Major file & folder restructuring. Split up several of the files into logical components and sorted them into relevenat folders.
Made some test adjustments to GigantismPlus' LevelText and Description.
