+++ ObjectBlueprints/Items.xml
@@ -314,7 +314,7 @@ You cannot wear equipment on your body."
     <part Name="CyberneticsGiganticExoframe" 
           Model="Alpha" 
           AugmentAdjectiveColor="b" 
-          AugmentTile="GiganticManipulator.png" 
+          AugmentTile="NaturalWeapons/GiganticManipulator.png" 
           AugmentTileColorString="&amp;c" 
           AugmentTileDetailColor="b" 
           AugmentedSwingSound="Sounds/Melee/cudgels/sfx_melee_cudgel_fullerite_swing"
