+++ Textures/Mutations/ElongatedPaws.png
