+++ Textures/NaturalWeapons/ElongatedBurrowingClaw.png
