+++ Textures/NaturalWeapons/ElongatedBurrowingCrystallinePoint.png
