+++ Textures/NaturalWeapons/ElongatedCrystallinePoint.png
