+++ Textures/NaturalWeapons/ElongatedPaw.png
