+++ Textures/NaturalWeapons/GiganticElongatedBurrowingClaw.png
