+++ Textures/NaturalWeapons/GiganticElongatedBurrowingCrystallinePoint.png
